Add WRITING.md file for contributors + about/meta/
Add a WRITING.md file as a guide for anyone that wants to
contribute to the site to get them up and running as fast
and smoothly as possible.

Add /about/meta/ section to the main site.
I'll host the same writing.md file there as well, including
assets like screenshots to illustrate how to install things.
The WRITING.md file will also reference these assets?