Update index.js for new packages
Updating all of the old Node packages.  Some things broke.

The biggest change was metalsmith-templates (deprecated).

Now we are using metalsmith-layouts and it requires some different
syntax in all of the files ("template:" -> "layout:").

It also required changing paths for metalsmith-pagination to work
properly, as well as a modification to the blog-list template.