Add Piwik analytics code
This adds the Piwiki analytics code provided (and hosted)
by andabata (Thank You andabata!).

Because last year Pat David was smart enough to use template
partials in handlebars, I simply had to add the Piwik code
to the `templates/partials/analytics.hbt` for it to work
everywhere else that GA previously did.

I also updated a few pages where there was no analytics code
because I'm lazy and forgot to add it back in the day.  So
yay for that! :)