Fix a bad blog post with multiple lede-img attr
This was causing the RSS feed to blow up and be very,
very unhappy. :(

In turn this was causing the PollFeed task on discuss to
spam the forum with a ton of old posts due to some wonky
crap in the rss/xml...  This should have fixed it.

The old posts will have been updated and should just stay
that way - they'll sink down with no activity on their own.