Consolidate index2.css + index.css in /
Thanks to prodding from @paperdigits this is to clean up
some spurious things.  /index2.css is only referenced in
index2-coming-soon.hbt template.

mv index2.css index.css
Fix template

I know this may get removed in the possible move to sass
in a while, but it can't hurt to clean a little? :)