Add <div>'s around fluid-vid content to render properly
I recently moved the site to a single-column layout, and part of
that migration broke a few things.  Fluid width video was one of
them.  I thought it was handled fine, but it picks up the
max-width of the containing column, so the video is wonky.

Wrapping all videos in <div> (or some block-level element) will
fix this.