articles: add legacy slugs from old site to all posts
This just uses the `slug` frontmatter key to set it to the same
slug that is on the legacy version of the site.
Hopefully this means old comment threads will get picked up with
the appropriate topic in discuss.

Don't bother manually specifying `slug` in future posts, it's not
needed.